Added support for nested JUnit5 tests.
As of version 1.2.0, only the first nested test class will be logged while the others are silently swallowed. Changing this required some adjustments to the new-line-handling which in turn required adjustments in other tests (effectively the first newline is now part of the suites block while the last one is now part of the summary block).